LPS-145491 Move CurrentAssignments from Assignments.js file
