+++ modules/dxp/apps/portal-workflow/portal-workflow-kaleo-designer-web/src/main/resources/META-INF/resources/designer/js/definition-builder/diagram-builder/components/sidebar/sections/assignments/Assignments.js
@@ -9,14 +9,13 @@
  * distribution rights of the Software.
  */
 
-import ClayButton, {ClayButtonWithIcon} from '@clayui/button';
-import ClayLayout from '@clayui/layout';
+import ClayButton from '@clayui/button';
 import PropTypes from 'prop-types';
 import React, {useContext} from 'react';
 
 import {DiagramBuilderContext} from '../../../../DiagramBuilderContext';
 import SidebarPanel from '../../SidebarPanel';
-import {options} from './SelectAssignment';
+import CurrentAssignments from './CurrentAssignments';
 
 const Assignments = ({setContentName}) => {
 	const {selectedItem} = useContext(DiagramBuilderContext);
@@ -33,51 +32,11 @@ const Assignments = ({setContentName}) => {
 				</ClayButton>
 			) : (
 				<CurrentAssignments
-					assignments={{...selectedItem.data.assignments}}
+				assignments = {selectedItem.data.assignments}
+				setContentName={setContentName}
 				/>
 			)}
 		</SidebarPanel>
-	);
-};
-
-const CurrentAssignments = ({assignments}) => {
-	const {setSelectedItem} = useContext(DiagramBuilderContext);
-	const deleteCurrentAssignments = () => {
-		setSelectedItem((previousValue) => ({
-			...previousValue,
-			data: {
-				...previousValue.data,
-				assignments: null,
-			},
-		}));
-	};
-
-	return (
-		<ClayLayout.ContentCol className="current-assignments-area" float>
-			<ClayLayout.Row
-				className="current-assignments-row"
-				justify="between"
-			>
-				<span>
-					{
-						options.find(
-							(option) =>
-								option.assignmentType ===
-								assignments.assignmentType[0]
-						)?.label
-					}
-
-					{/* {Liferay.Language.get(assignments.assignmentType[0])} */}
-				</span>
-
-				<ClayButtonWithIcon
-					className="delete-button text-secondary trash-button"
-					displayType="unstyled"
-					onClick={deleteCurrentAssignments}
-					symbol="trash"
-				/>
-			</ClayLayout.Row>
-		</ClayLayout.ContentCol>
 	);
 };
